type/mock cleaning
