+++ frontend/src/commons/mock/MockStudent.ts
@@ -1,0 +1,45 @@
+
+import { Student } from "./../types/TypeStudent";
+
+export const MockStudent: Student = {
+    metadata: [
+        {
+            grade: "First-Year",
+            calendarYear: "2022-2023",
+            fall: {
+                courses: [
+                    { course: { studentCourse: true, code: "DRST 001", title: "Directed Studies: Literature", seasons: ["FALL", "SPRING"], evaluation: { rating: 4.1, workload: 4.2, professor: 5}, distribution: ["Hu", "WR"] }, enrollmentStatus: "COMPLETED", season: "FALL", year: "2022-2023", letterGrade: "A-" }, 
+                    { course: { studentCourse: true, code: "DRST 001", title: "Directed Studies: Literature", seasons: ["FALL", "SPRING"], evaluation: { rating: 4.1, workload: 4.2, professor: 5}, distribution: ["Hu", "WR"] }, enrollmentStatus: "COMPLETED", season: "FALL", year: "2022-2023", letterGrade: "A-" }, 
+                    { course: { studentCourse: true, code: "DRST 001", title: "Directed Studies: Literature", seasons: ["FALL", "SPRING"], evaluation: { rating: 4.1, workload: 4.2, professor: 5}, distribution: ["Hu", "WR"] }, enrollmentStatus: "COMPLETED", season: "FALL", year: "2022-2023", letterGrade: "A-" }, 
+                    { course: { studentCourse: true, code: "DRST 001", title: "Directed Studies: Literature", seasons: ["FALL", "SPRING"], evaluation: { rating: 4.1, workload: 4.2, professor: 5}, distribution: ["Hu", "WR"] }, enrollmentStatus: "COMPLETED", season: "FALL", year: "2022-2023", letterGrade: "A-" },]
+            },
+            spring: {
+                courses: [
+                    { course: { studentCourse: true, code: "DRST 001", title: "Directed Studies: Literature", seasons: ["FALL", "SPRING"], evaluation: { rating: 4.1, workload: 4.2, professor: 5}, distribution: ["Hu", "WR"] }, enrollmentStatus: "COMPLETED", season: "FALL", year: "2022-2023", letterGrade: "A-" }, 
+                    { course: { studentCourse: true, code: "DRST 001", title: "Directed Studies: Literature", seasons: ["FALL", "SPRING"], evaluation: { rating: 4.1, workload: 4.2, professor: 5}, distribution: ["Hu", "WR"] }, enrollmentStatus: "COMPLETED", season: "FALL", year: "2022-2023", letterGrade: "A-" }, 
+                    { course: { studentCourse: true, code: "DRST 001", title: "Directed Studies: Literature", seasons: ["FALL", "SPRING"], evaluation: { rating: 4.1, workload: 4.2, professor: 5}, distribution: ["Hu", "WR"] }, enrollmentStatus: "COMPLETED", season: "FALL", year: "2022-2023", letterGrade: "A-" }, 
+                    { course: { studentCourse: true, code: "DRST 001", title: "Directed Studies: Literature", seasons: ["FALL", "SPRING"], evaluation: { rating: 4.1, workload: 4.2, professor: 5}, distribution: ["Hu", "WR"] }, enrollmentStatus: "COMPLETED", season: "FALL", year: "2022-2023", letterGrade: "A-" }, 
+                    { course: { studentCourse: true, code: "DRST 001", title: "Directed Studies: Literature", seasons: ["FALL", "SPRING"], evaluation: { rating: 4.1, workload: 4.2, professor: 5}, distribution: ["Hu", "WR"] }, enrollmentStatus: "COMPLETED", season: "FALL", year: "2022-2023", letterGrade: "A-" },]
+            }
+        },
+        {
+            grade: "Sophomore",
+            calendarYear: "2023-2024",
+            fall: {
+                courses: [
+                    { course: { studentCourse: true, code: "DRST 001", title: "Directed Studies: Literature", seasons: ["FALL", "SPRING"], evaluation: { rating: 4.1, workload: 4.2, professor: 5}, distribution: ["Hu", "WR"] }, enrollmentStatus: "COMPLETED", season: "FALL", year: "2022-2023", letterGrade: "A-" }, 
+                    { course: { studentCourse: true, code: "DRST 001", title: "Directed Studies: Literature", seasons: ["FALL", "SPRING"], evaluation: { rating: 4.1, workload: 4.2, professor: 5}, distribution: ["Hu", "WR"] }, enrollmentStatus: "COMPLETED", season: "FALL", year: "2022-2023", letterGrade: "A-" }, 
+                    { course: { studentCourse: true, code: "DRST 001", title: "Directed Studies: Literature", seasons: ["FALL", "SPRING"], evaluation: { rating: 4.1, workload: 4.2, professor: 5}, distribution: ["Hu", "WR"] }, enrollmentStatus: "COMPLETED", season: "FALL", year: "2022-2023", letterGrade: "A-" }, 
+                    { course: { studentCourse: true, code: "DRST 001", title: "Directed Studies: Literature", seasons: ["FALL", "SPRING"], evaluation: { rating: 4.1, workload: 4.2, professor: 5}, distribution: ["Hu", "WR"] }, enrollmentStatus: "COMPLETED", season: "FALL", year: "2022-2023", letterGrade: "A-" }, 
+                    { course: { studentCourse: true, code: "DRST 001", title: "Directed Studies: Literature", seasons: ["FALL", "SPRING"], evaluation: { rating: 4.1, workload: 4.2, professor: 5}, distribution: ["Hu", "WR"] }, enrollmentStatus: "COMPLETED", season: "FALL", year: "2022-2023", letterGrade: "A-" },] 
+            },
+            spring: {
+                courses: [
+                    { course: { studentCourse: true, code: "DRST 001", title: "Directed Studies: Literature", seasons: ["FALL", "SPRING"], evaluation: { rating: 4.1, workload: 4.2, professor: 5}, distribution: ["Hu", "WR"] }, enrollmentStatus: "COMPLETED", season: "FALL", year: "2022-2023", letterGrade: "A-" }, 
+                    { course: { studentCourse: true, code: "DRST 001", title: "Directed Studies: Literature", seasons: ["FALL", "SPRING"], evaluation: { rating: 4.1, workload: 4.2, professor: 5}, distribution: ["Hu", "WR"] }, enrollmentStatus: "COMPLETED", season: "FALL", year: "2022-2023", letterGrade: "A-" }, 
+                    { course: { studentCourse: true, code: "DRST 001", title: "Directed Studies: Literature", seasons: ["FALL", "SPRING"], evaluation: { rating: 4.1, workload: 4.2, professor: 5}, distribution: ["Hu", "WR"] }, enrollmentStatus: "COMPLETED", season: "FALL", year: "2022-2023", letterGrade: "A-" }, 
+                    { course: { studentCourse: true, code: "DRST 001", title: "Directed Studies: Literature", seasons: ["FALL", "SPRING"], evaluation: { rating: 4.1, workload: 4.2, professor: 5}, distribution: ["Hu", "WR"] }, enrollmentStatus: "COMPLETED", season: "FALL", year: "2022-2023", letterGrade: "A-" },] 
+            }
+        }
+    ]
+};
