+++ frontend/src/commons/types/TypeStudent.ts
@@ -1,0 +1,16 @@
+
+import { StudentCourse } from "./TypeCourse";
+
+export interface Semester {
+    courses: Array<StudentCourse>;
+}
+export interface Year {
+    grade: "First-Year" | "Sophomore" | "Junior" | "Senior";
+    calendarYear: string;
+    fall : Semester;
+    spring: Semester;
+}
+
+export interface Student {
+    metadata: Array<Year>;
+}
